back-end/requirements.txt
blinker==1.9.0
click==8.3.1
cs50==9.3.4
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.2
packaging==24.1
pandas==3.0.0
pillow==10.3.0
psycopg2-binary==2.9.11
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
qrcode==7.4.2
six==1.17.0
SQLAlchemy==2.0.31
sqlparse==0.5.0
termcolor==2.4.0
typing_extensions==4.12.2
Werkzeug==3.1.5
